Add plug-in based design blueprint samples
Aria Automation introduces new plug-in based designs and
deployments model in 8.12.0 (April) release. We have a
tutorial to help users get started with the new
capabilities. The tutorial includes 6 versions of a
plug-in based design blueprint.

Add all versions of the blueprint with descriptions to the
cloud assembly blueprint samples to allow users to easily
take them, tweak them for their use cases and evolve them
however they need.

Signed-off-by: Kristina Zhelyazkova <kzhelyazkova@vmware.com>